Add functional tests for add_targets + small improvements (#169)
* Fix http_request() fixture typos

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Add additional permissions in access_token scope

An additional permissions are required in order to succesfully test
the add targets feature.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Changes to successful case in add_targets feature

Some changes had to be made in the successful scenario called
"Adding a target using RSTUF API" in add_targets. feature file in order
to achieve reproducibility in the code and make sure steps can be
easily implemented.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Drop redundant scenarios from add_targets feature

I removed in total three scenarios:
1) adding the same target twice using RSTUF API
2) adding a target with insufficient token scope using RSTUF API
3) adding a target with invalid information using RSTUF API

The first one I removed as adding a target twice doesn't test any
logic at all. The behavior of RSTUF is such that no matter how many
duplicate targets you will add synchronously or asynchronously it will
always add them separately and bump the version of the corresponding target.

The second and third one I found that are already tested in unit
tests inside RSTUF API. See:
https://github.com/vmware/repository-service-tuf-api/blob/main/tests/unit/api/test_targets.py

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Add add_targets.feature functional tests

It was required to additionally add a root.json example file
and get_target_info() pytest.feature in conftest.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Fix formatting errors and use other root location

Use root.json that's located inside a local "metadata" folder instead
the one I have manually stored in "tests/data" because each time
we do a bootstrap the expiry is different as expiry is based on
the current time.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* functional workflow: copy metadata folder as well

Copy the content of the "metadata" content inside the umbrella
repository, so that ngclient can use "root.json".

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Address Kairo's comments

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>