Rename one of the remove targets scenario examples
Renaming one of the examples in the remove targets feature file was
necessarily because of a bug I noticed in how our publishing works when
there is a delete operation requested.
See https://github.com/vmware/repository-service-tuf-worker/issues/152
Because of this issue, the timeout was hit over and over again and
in order to escape the situation when one bin is responsible for two
targets I just renamed at the same time I just renamed one of the
targets.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>