Remove the "bin" role from diagrams and update make docs command (#232)
* Update "make docs" plantuml command

Update the "plantuml" command in the "make docs" so that the
"docs/source/_static/1_1_rstuf_metadata.png" image file is updated
when changes are made inside "docs/diagrams/1_1_rstuf_metadata.puml".

The problem was that the old regex implied that the names of the images
will finish exactly with "rstuf.png", but the "1_1_rstuf_metadata.puml"
file name didn't respect that rule and the rule had to be updated.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Remove the "bin" role from diagrams

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

---------

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>