New RSTUF Guide
The guide is rendered in https://repository-service-tuf.readthedocs.io/en/v1.0.0a1-draft/

It changes the Guide structure entirely to give users how and when to
use TUF end-to-end with RSTUF. It starts with a general introduction,
presenting RSTUF and having a 50s video on which kind of protection
we aim for, and even gives a short What TUF is, and we provide some
examples of use cases.

After this lightweight-top-level introduction, we go to how you can
use RSTUF. We show the simple API basic usage (Using RSTUF) --  the
easy integration/usage is the core RSTUF's strength. We talk about
Signing Keys. We go for the Deployment part, which now shows guides
to deploy using Docker and Kubernetes, also have in-depth Deployment
Planning for more customized deployments (likewise advanced support
for Docker and Kubernetes deployment), finishing this with the
Configuration & Setup, talking about the Ceremony and Bootstrap.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>