Functional tests uses `RSTUF CLI` with `--auth`
As implemented by CLI
(https://github.com/repository-service-tuf/repository-service-tuf-cli/pull/289)
the functional tests require to use the `--auth` parameters to be
compatible with the existent tests, these tests use a API running the
built-in authentication.

Note: We are planning changing the FT to be compatible with issue #41
This commit doesn't aim this change. It fix the current state.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>