New RSTUF Guide (#341)
* New RSTUF Guide

The guide is rendered in https://repository-service-tuf.readthedocs.io/en/v1.0.0a1-draft/

It changes the Guide structure entirely to give users how and when to
use TUF end-to-end with RSTUF. It starts with a general introduction,
presenting RSTUF and having a 50s video on which kind of protection
we aim for, and even gives a short What TUF is, and we provide some
examples of use cases.

After this lightweight-top-level introduction, we go to how you can
use RSTUF. We show the simple API basic usage (Using RSTUF) --  the
easy integration/usage is the core RSTUF's strength. We talk about
Signing Keys. We go for the Deployment part, which now shows guides
to deploy using Docker and Kubernetes, also have in-depth Deployment
Planning for more customized deployments (likewise advanced support
for Docker and Kubernetes deployment), finishing this with the
Configuration & Setup, talking about the Ceremony and Bootstrap.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* docs: update clients section (#357)

* New RSTUF Guide

The guide is rendered in https://repository-service-tuf.readthedocs.io/en/v1.0.0a1-draft/

It changes the Guide structure entirely to give users how and when to
use TUF end-to-end with RSTUF. It starts with a general introduction,
presenting RSTUF and having a 50s video on which kind of protection
we aim for, and even gives a short What TUF is, and we provide some
examples of use cases.

After this lightweight-top-level introduction, we go to how you can
use RSTUF. We show the simple API basic usage (Using RSTUF) --  the
easy integration/usage is the core RSTUF's strength. We talk about
Signing Keys. We go for the Deployment part, which now shows guides
to deploy using Docker and Kubernetes, also have in-depth Deployment
Planning for more customized deployments (likewise advanced support
for Docker and Kubernetes deployment), finishing this with the
Configuration & Setup, talking about the Ceremony and Bootstrap.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* docs: update clients section

* emphasize how rstuf takes care of repo side
* promote python-tuf ngclient
* note that other libraries may not be fully compatible
* add section about trust bootstrap
* remove detail of metadata update

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

---------

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>
Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>
Co-authored-by: Kairo de Araujo <kdearaujo@vmware.com>

* Add small diagram to TUF client

Add small diagram to TUF client ilustrating the layer.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Apply suggestions from @rhondae's code review

Adds @rhondae's code review suggestions for grammar, wording, and clarity in the new Guide.

Co-authored-by: rhondae <115107029+rhondae@users.noreply.github.com>

* Add suggestions from code review (English/Wording)

Address comments from @rhondae's review

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Fix typo in TUF

TUF typo (as TU)

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Apply @rhondae suggestions from code review

Add the amazing @rhondae suggestion to RSTUF Guide
This fixes English, grammar, and wording.

Co-authored-by: rhondae <115107029+rhondae@users.noreply.github.com>

* Add fixes from comments: signing and generated

- Fix misspelled Singing to Signing
- Fix wording for to skip online keys when 'generated'

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Add sphinx-toolbox as dependency

We use the `'sphinx_toolbox.collapse'` to some examples

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Add new RSTUF Logo and Icon

Add the new RSTUF Logo and Icon

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Fix logo paths

- README.rst
- Guide index.rst

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Fix requirements with pipenv 2023.7.9

Add updates for new Pipenv for generating the requirements

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Remove deprecated `--no-auth` CLI parameter

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Add the Docker and Kubernetes to use RSTUF latest

Add the Docker and Kubernetes to use RSTUF latest images

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

* Add extra details to start the Docker RSTUF deployment

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

---------

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>
Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>
Co-authored-by: Lukas Pühringer <luk.puehringer@gmail.com>
Co-authored-by: rhondae <115107029+rhondae@users.noreply.github.com>