Add GHA env variables for RSTUF setup scripts
The FT uses now the RSTUF setup scripts for running RSTUF setup during
the ceremony and signing for example.

It required to run the scripts from `rstuf-umbrella/tests/` when
triggered by another component repository.

These are part of setting up the RSTUF deployment for running the
functional tests.

This commit also adds a step to get the trusted tuf from the TUF
repository metadata

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>