Support split test run (#603)
* Use `pytest-split`

* Split functional tests

- We use `pytest-split` to split the tests
and run them parallely

- Since we don't store the test durations yet,
`pytest-split` does not choose the ideal splitting
groups. Thus, we run the very slow tests separately
by providing a parameter whether to run or not the
SLOW tests.

* Adapt Functional Tests CI to splitted tests

* docs: Add `pytest-xdist` to dev docs

* docs: Add `pytest-split` to dev docs

* Explain `splits` value in `pytest-split`

* Fix requirements check failing

The check in linting with the diff tool between
the requirements files and the pipenv requirements
commands fails because of an extra space added after
the ';' character, e.g.:

2c2
< certifi==2022.9.24; python_full_version >= '3.6.0'
---
> certifi==2022.9.24 ; python_full_version >= '3.6.0'

We add the -w option (--ignore-all-space) to resolve
this issue with unknown yet root cause.