Make parameters unique for ft-signed params
Duplicates in run-ft-signed paramters for the script
"rstuf-admin-ceremony.py" are removed.
For example if we have:
s = '{
    "Please enter path to encrypted private key": "tests/.../JJ.ecdsa",
    "Please enter password": "hunter2",
    "Please enter signing key index, or press enter to continue": "1",
    "Please enter path to encrypted private key": "tests/.../JH.ed25519",
    "Please enter password": "hunter2"
}'

and do:

import json
d = json.loads(s)
print(d)

the result will be:

{
    "Please enter path to encrypted private key": "tests/.../JJ.ecdsa",
    "Please enter password": "hunter2",
    "Please enter signing key index, or press enter to continue": "1",
}

meaning the answers for the second root key will be removed as there
are dict items with the same keys.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>