Fix ft tests
With the change in how we do metadata updates, we now fetch the latest
root version.
Before we gave a URL to a specific root in this case root version 2,
but actually, this was not the latest root and after we changed the
the way we do "metadata update" the tests are failing.

Signed-off-by: Martin Vrachev <martin.vrachev@gmail.com>