tests: FT update + DAS + sign
This commit implements to the FT the test for updating the root
metadata with a not fully signed metadata.

The changes are made first in the `ft-das` and in the `Metdata Update`
FT.
It starts the DAS Bootstrap and generates also a Root Metadata Update
with DAS.
The FT  updates the Root Metadata without all required (thresold
signatures).
While doing the signing it keep sending artifacts to the TUF metadata
and we validate the consistency of the metadata.

Signed-off-by: Kairo Araujo <kairo.araujo@testifysec.com>